// Angular/unused/tournament/models/tournament-registration.ts
export interface TournamentRegistrationApplication {
  tournamentId: number;
  teamName: string;
  captainEmail: string;
  battletag1: string;
  battletag2: string;
  battletag3: string;
  battletag4: string;
  battletag5: string;
  battletag6: string;
  battletag7: string;
  paypalEmail: string;
  isPaid: number;
}